Move tests for PUT /ctms/{email_id} (create or update)
